third_party/WebKit/LayoutTests/http/tests/serviceworker/chromium/navigation-preload-origin-trial-methods-expected.txt

CONSOLE WARNING: line 13: This test relies on console message comparison so there can be different -expected.txt files for virtual test suites.
CONSOLE MESSAGE: line 68: --normal SW--
disable() rejected with: AbortError: Failed to enable or disable navigation preload: The service worker script does not have a valid Navigation Preload Origin Trial token.
enable() rejected with: AbortError: Failed to enable or disable navigation preload: The service worker script does not have a valid Navigation Preload Origin Trial token.
getState() resolved with: {"enabled":false,"headerValue":"true"}
setHeaderValue() rejected with: AbortError: Failed to set navigation preload header: The service worker script does not have a valid Navigation Preload Origin Trial token.
log from SW
Activate event handler
registration.navigationPreload not found
Fetch event handler
event.preloadResponse not found

CONSOLE MESSAGE: line 79: --Origin-Trial enabled SW--
disable() resolved with: undefined
enable() resolved with: undefined
getState() resolved with: {"enabled":true,"headerValue":"world"}
setHeaderValue() resolved with: undefined
log from SW
Activate event handler
disable() resolved with: undefined
enable() resolved with: undefined
getState() resolved with: {"enabled":true,"headerValue":"true"}
setHeaderValue() resolved with: undefined
Fetch event handler
event.preloadResponse resolved with: [object Response]
Text of preloadResponse: [Service-Worker-Navigation-Preload header set: hello]

This is a testharness.js-based test.
PASS Calling Navigation preload related methods for normal SW. 
PASS Calling Navigation preload related methods for Origin-Trial enabled SW. 
Harness: the test ran to completion.
